requirements.txt
# RAG Document Q&A Chatbot Requirements
# Python >= 3.10 required
# Install with: pip install -r requirements.txt

# Core Framework
streamlit==1.33.0

# LangChain & LLM Integration
langchain==0.2.3
langchain-core==0.2.43
langchain-community==0.2.4
langchain-groq==0.1.4

# Embeddings & Vector Storage
sentence-transformers==2.6.1
faiss-cpu==1.7.4
torch==2.2.2
transformers==4.38.2
tokenizers==0.15.2

# Document Processing
pypdf==4.2.0

# Environment & Utilities
python-dotenv==1.0.1
rich==13.7.1